+++ ConcernsCaseWork/release-notes.md
@@ -1,5 +1,6 @@
 ## 19.0.0
 
+* Added Companies House Number to concerns cases. Each new concerns case will store the Companies House Number of the associated Trust (existing cases will be updated in retrospect).
 [To be completed]
 
 ---